fix: Menu tests
